+++ app/grandchallenge/evaluation/admin.py
@@ -25,7 +25,6 @@ from grandchallenge.evaluation.models import (
     EvaluationGroundTruthUserObjectPermission,
     EvaluationGroupObjectPermission,
     EvaluationUserObjectPermission,
-    EvaluationUtilization,
     Method,
     MethodGroupObjectPermission,
     MethodUserObjectPermission,
@@ -215,30 +214,6 @@ class EvaluationAdmin(admin.ModelAdmin):
         "evaluation_utilization",
     )
     actions = (requeue_jobs, cancel_jobs, deprovision_jobs)
-
-
-@admin.register(EvaluationUtilization)
-class EvaluationUtilizationAdmin(admin.ModelAdmin):
-    ordering = ("-created",)
-    list_display = (
-        "pk",
-        "created",
-        "evaluation",
-        "creator",
-        "duration",
-        "compute_cost_euro_millicents",
-        "phase",
-        "challenge",
-        "algorithm_image",
-        "algorithm",
-    )
-    search_fields = (
-        "creator__username",
-        "pk",
-        "evaluation__pk",
-        "phase__slug",
-        "challenge__short_name",
-    )
 
 
 @admin.register(CombinedLeaderboard)
